% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qc.test}
\alias{qc.test}
\title{Kwaliteitsanalyse o.b.v. Nelson, Westgard of anders}
\usage{
qc.test(x, dates = rep(NA, length(x)), m = mean(x), s = sd(x),
  round = 2, type = "Nelson", text.show = TRUE, plot.show = TRUE,
  colours = rainbow(8), include_EWMA = TRUE,
  title = paste("Quality Control according to", type), subtitle = "",
  subtitle.colour = colours[1], plot.withoutrule1 = FALSE,
  plot.print = TRUE, x.lbl = "Observation", y.lbl = "Value",
  markdown = FALSE, force = FALSE)
}
\arguments{
\item{x}{Gegevensreeks.}

\item{dates}{Standaard is \code{rep(NA, length(x))}. Datums van gegevenswaarden.}

\item{m}{Standaard is \code{mean(x)}. Gemiddelde van \code{x} dat gebruikt wordt in de kwaliteitsregels.}

\item{s}{Standaard is \code{sd(x)}. Standaardafwijking van \code{x} dat gebruikt wordt in de kwaliteitsregels.}

\item{round}{Standaard is \code{2}. Aantal decimalen waarop alle getallen afgerond worden.}

\item{type}{Standaard is \code{"Nelson"}. Geldige opties zijn \code{"Nelson"}, \code{"Westgard"}, \code{"AIAG"}, \code{"Montgomery"} of \code{"Healthcare"}. Zie \strong{Rules list} (onderaan) voor meer informatie.}

\item{text.show}{Standaard is \code{TRUE}. Tekstanalyse weergeven.}

\item{plot.show}{Standaard is \code{TRUE}. Grafiek weergeven met gemiddelde (zie \code{\link[certedata]{mean}}), EWMA (zie \code{\link{ewma}}), 1-3x de standaardafwijking (zie \code{\link[certedata]{sd}}) en de waarnemingen met eventueel overtreden kwaliteitsregels.}

\item{title}{Standaard is \code{paste("Kwaliteitscontrole volgens", type)}. Titel van de toets en de grafiek. Tekst tussen sterretjes wordt cursief gemaakt.}

\item{subtitle}{Standaard is \code{""}. Ondertitel van de grafiek. Tekst tussen sterretjes wordt cursief gemaakt.}

\item{subtitle.colour}{Standaard is \code{colourpicker("certeblauw")}. Zie ook \code{\link{colourpicker}}. Kleur van de ondertitel.}

\item{plot.withoutrule1}{Standaard is \code{FALSE}. Als regel 1 overtreden wordt (zie \code{\link{rule1}}), de plot opnieuw weergeven met data zonder deze waarnemingen.}

\item{plot.print}{Standaard is \code{TRUE}. De grafiek direct printen met de functie \code{print}. Met \code{FALSE} wordt de grafiek als ggplot-model geretourneerd.}

\item{x.lbl}{Standaard is \code{"Waarneming"}. De tekst op de x-as.}

\item{y.lbl}{Standaard is \code{"Waarde"}. De tekst op de y-as.}

\item{markdown}{Standaard is \code{FALSE}. Tekstanalyse afdrukken in markdown-formaat met \code{\link{tbl_markdown}}.}

\item{force}{Standaard is \code{FALSE}. Bij grote afwijkingen wordt de analyse gestaakt. Gebruik deze optie om de analyse te forceren.}
}
\description{
Met deze functie kan een gegevensreeks getoetst worden op kwaliteitsregels. Daarnaast wordt getoetst of er een significante lineaire trend bestaat.
}
\keyword{nelson}
\keyword{qc}
\keyword{qcc}
\keyword{westgard}
