% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\alias{rule1}
\alias{rule}
\alias{nelson}
\alias{rule2}
\alias{rule3}
\alias{rule4}
\alias{rule5}
\alias{rule6}
\alias{rule7}
\alias{rule8}
\title{Nelson's Quality Control Rules}
\source{
Nelson LS. \strong{The Shewhart Control Chart—Tests for Special Causes}. Journal of Quality Technology [Internet]. Informa UK Limited; 1984 Oct;16(4):237–9. Available from: \url{http://dx.doi.org/10.1080/00224065.1984.11978921}
}
\usage{
rule1(x, m = mean(x), s = sd(x))

rule2(x, m = mean(x), min.run = 9)

rule3(x, min.run = 6)

rule4(x, m = mean(x), min.run = 14, direction.mean = FALSE)

rule5(x, m = mean(x), s = sd(x), min.run = 3)

rule6(x, m = mean(x), s = sd(x), min.run = 5)

rule7(x, m = mean(x), s = sd(x), min.run = 15)

rule8(x, m = mean(x), s = sd(x), min.run = 8)
}
\arguments{
\item{x}{vector with values}

\item{m}{mean}

\item{s}{standard deviation}

\item{min.run}{minimal amount of sequential values before rule is triggered (defaults to Nelson's)}

\item{direction.mean}{a logical to indicate whether \emph{n} observations in a row must be tested for alternating in direction of the mean}
}
\description{
These rules are used for quality control (QC). Default values are set for Nelson's criteria, but they also support Westgard, AIAG, Montgomery and Healthcare QC rules.
}
\section{Rules list}{

  \tabular{lccccc}{
    \emph{Nelson (N), Westgard (W), AIAG (A), Montgomery (M), Healthcare (H):}             \tab  N \tab  W \tab  A \tab  M \tab  H\cr
    ---------------------------------------------------------------------------------------\tab ----- \tab ----- \tab ----- \tab  ----- \tab -----\cr
    \strong{#1} One point is more than 3 standard deviations from the mean.                \tab  1 \tab  1 \tab  1 \tab  1 \tab  1\cr
    \strong{#2} \emph{n} (or more) points in a row are on the same side of the mean        \tab  9 \tab  9 \tab  7 \tab  8 \tab  8\cr
    \strong{#3} \emph{n} (or more) points in a row are continually incr. or decr.          \tab  6 \tab  - \tab  6 \tab  6 \tab  6\cr
    \strong{#4} \emph{n} (or more) points in a row alternate in direction, incr. then decr.\tab 14 \tab -  \tab 14 \tab 14 \tab - \cr
    \strong{#5} \emph{n} - 1 out of \emph{n} points in a row are >2 sd from the mean       \tab  3 \tab  3 \tab  3 \tab  3 \tab  3\cr
    \strong{#6} \emph{n} - 1 out of \emph{n} points in a row are >1 sd from the mean       \tab  5 \tab  5 \tab  5 \tab  5 \tab - \cr
    \strong{#7} >= \emph{n} points in a row are within 1 sd of the mean                    \tab 15 \tab  - \tab 15 \tab 15 \tab 15\cr
    \strong{#8} >= \emph{n} points in a row outside 1 sd of the mean, in both directions   \tab  8 \tab  - \tab  8 \tab  8 \tab -
  }
}

\keyword{nelson}
\keyword{qc}
\keyword{qcc}
\keyword{rule}
